requirements.txt
jinja2
Alfred-Workflow
boto3
click
